Loading only one widget js file
