+++ unitylibs/core/widgets/prompt-with-style-select/prompt-with-style-select.js
@@ -1,4 +1,5 @@
 /* eslint-disable class-methods-use-this */
+/* eslint-disable max-classes-per-file -- UnityWidget is inlined with PromptWithStyleSelectWidget for a single network fetch */
 
 /**
  * Parses Unity block authoring for prompt + style selection (Unity block with .widget-prompt-with-style).
@@ -9,13 +10,452 @@
  * each row has three column `<div>`s (mobile / tablet / desktop) with a `<picture>` each.
  */
 
-import UnityWidget from '../../workflow/workflow-firefly/widget.js';
 import {
   createTag,
   defineDeviceByScreenSize,
   getUnityLibs,
-  loadStyle,
 } from '../../../scripts/utils.js';
+
+/** Duplicated UnityWidget — keep aligned with `prompt-widget/prompt-widget.js`. */
+/**
+ * Shared Firefly widget: verb picker ({@link #verbDropdown}), model picker, Generate button,
+ * and dropdown plumbing for {@link PromptWidget} (hero) and {@link PromptWithStyleSelectWidget}.
+ */
+export class UnityWidget {
+  constructor(target, el, workflowCfg, spriteCon) {
+    this.el = el;
+    this.target = target;
+    this.workflowCfg = workflowCfg;
+    this.widget = null;
+    this.actionMap = {};
+    this.spriteCon = spriteCon;
+    this.prompts = null;
+    this.models = null;
+    this.selectedVerbType = '';
+    this.selectedVerbText = '';
+    this.selectedModelModule = '';
+    this.selectedModelId = '';
+    this.selectedModelText = '';
+    this.selectedModelVersion = '';
+    this.promptItems = [];
+    this.genBtn = null;
+    this.hasPromptSuggestions = false;
+    this.hasModelOptions = false;
+    this.lanaOptions = { sampleRate: 100, tags: 'Unity-FF' };
+    this.sound = { audio: null, currentTile: null, currentUrl: '' };
+    this.durationCache = new Map();
+  }
+
+  /**
+   * Default for Firefly: marquee prompt bar via {@link initPromptWidget}.
+   * Class `widget-prompt-with-style` uses {@link PromptWithStyleSelectWidget} instead (see workflow.js).
+   */
+  async initWidget() {
+    const widgetsBase = `${getUnityLibs()}/core/widgets`;
+    const { initPromptWidget } = await import(`${widgetsBase}/prompt-widget/prompt-widget.js`);
+    return initPromptWidget(this);
+  }
+
+  /** No-op on base; marquee subclass loads sound augmentation for the `sound` verb. */
+  async ensureSoundModuleLoaded() {
+    await Promise.resolve();
+  }
+
+  /**
+   * Media-type combobox from authored `[class*="icon-verb"]` + sibling link text.
+   * Used by the hero prompt bar and prompt-with-style-select.
+   *
+   * @returns {HTMLElement[]} `[selectedButton]` when a single verb, else `[selectedButton, verbListUl]`
+   */
+  verbDropdown() {
+    const verbs = this.el.querySelectorAll('[class*="icon-verb"]');
+    const inputPlaceHolder = this.el.querySelector('.icon-placeholder-input').parentElement.textContent;
+    const selectedVerbType = verbs[0]?.className.split('-')[2];
+    const selectedVerb = verbs[0]?.nextElementSibling;
+    const selectedElement = createTag('button', {
+      class: 'selected-verb',
+      'aria-expanded': 'false',
+      'aria-controls': 'media-menu',
+      'aria-label': 'media type',
+      'aria-haspopup': 'listbox',
+      role: 'combobox',
+      'aria-labelledby': 'listbox-label',
+      'data-selected-verb': selectedVerbType,
+    }, `${selectedVerb?.textContent.trim()}`);
+    this.selectedVerbType = selectedVerbType;
+    this.widgetWrap.setAttribute('data-selected-verb', this.selectedVerbType);
+    this.selectedVerbText = selectedVerb?.textContent.trim();
+    if (verbs.length <= 1) {
+      selectedElement.setAttribute('disabled', 'true');
+      return [selectedElement];
+    }
+    this.widgetWrap.classList.add('verb-options');
+    const menuIcon = createTag('span', { class: 'menu-icon' }, '<svg><use xlink:href="#unity-chevron-icon"></use></svg>');
+    const verbList = createTag('ul', { class: 'verb-list', id: 'media-menu', role: 'listbox', 'aria-labelledby': 'listbox-label' });
+    verbList.setAttribute('style', 'display: none;');
+    selectedElement.append(menuIcon);
+    const handleDocumentClick = (e) => {
+      const menuContainer = selectedElement.parentElement;
+      if (!menuContainer.contains(e.target)) {
+        document.removeEventListener('click', handleDocumentClick);
+        this.closeVerbOrModelMenu(selectedElement);
+      }
+    };
+    selectedElement.addEventListener('click', (e) => {
+      e.stopPropagation();
+      this.hidePromptDropdown(selectedElement);
+      this.showVerbMenu(selectedElement);
+      document.addEventListener('click', handleDocumentClick);
+    }, true);
+    selectedElement.addEventListener('keydown', (e) => {
+      if (e.key === 'Enter' || e.key === ' ') {
+        this.hidePromptDropdown(selectedElement);
+        this.showVerbMenu(selectedElement);
+      }
+      if (e.key === 'Escape' || e.code === 27) {
+        this.closeVerbOrModelMenu(selectedElement);
+        selectedElement.focus();
+      }
+    });
+    verbs[0]?.classList.add('selected');
+    const verbsData = Array.from(verbs).map((verb) => ({
+      name: verb.nextElementSibling?.textContent.trim(),
+      type: verb.classList[1].split('-')[2],
+      icon: verb.nextElementSibling?.href,
+    }));
+    this.createDropdownItems(verbsData, verbList, selectedElement, menuIcon, inputPlaceHolder, false);
+    return [selectedElement, verbList];
+  }
+
+  /**
+   * Collapses a verb or model combobox. Restores inline `display: none` on the `.verb-list`
+   * (opening removes it in `showVerbMenu`; without this the list stays visible when
+   * `.show-menu` is cleared — `display: block` only applies while `.show-menu` is set).
+   *
+   * @param {HTMLElement} selectedElement — `.selected-verb` or `.selected-model`
+   */
+  closeVerbOrModelMenu(selectedElement) {
+    const menuContainer = selectedElement?.parentElement;
+    if (!menuContainer) return;
+    menuContainer.classList.remove('show-menu');
+    selectedElement.setAttribute('aria-expanded', 'false');
+    const list = selectedElement.nextElementSibling;
+    if (list?.classList?.contains('verb-list')) {
+      list.setAttribute('style', 'display: none;');
+    }
+  }
+
+  showVerbMenu(selectedElement) {
+    const menuContainer = selectedElement.parentElement;
+    document.querySelectorAll('.verbs-container').forEach((container) => {
+      if (container !== menuContainer) {
+        const sv = container.querySelector('.selected-verb');
+        if (sv) this.closeVerbOrModelMenu(sv);
+      }
+    });
+    document.querySelectorAll('.models-container').forEach((container) => {
+      if (container !== menuContainer) {
+        const sm = container.querySelector('.selected-model');
+        if (sm) this.closeVerbOrModelMenu(sm);
+      }
+    });
+    menuContainer.classList.toggle('show-menu');
+    const isOpen = menuContainer.classList.contains('show-menu');
+    selectedElement.setAttribute('aria-expanded', isOpen ? 'true' : 'false');
+    const siblingList = selectedElement.nextElementSibling;
+    if (siblingList?.classList?.contains('verb-list')) {
+      if (isOpen) {
+        siblingList.removeAttribute('style');
+      } else {
+        siblingList.setAttribute('style', 'display: none;');
+      }
+    }
+  }
+
+  hidePromptDropdown(exceptElement = null) {
+    const dropdown = this.widget.querySelector('.prompt-dropdown-container');
+    if (dropdown && !dropdown.classList.contains('hidden')) {
+      dropdown.classList.add('hidden');
+      dropdown.setAttribute('inert', '');
+      dropdown.setAttribute('aria-hidden', 'true');
+    }
+    if (this.selectedVerbType === 'sound') {
+      this.resetAllSoundVariations?.(dropdown);
+    }
+    const modelDropdown = this.widget.querySelector('.models-container');
+    const modelButton = modelDropdown?.querySelector('.selected-model');
+    if (modelDropdown && modelDropdown.classList.contains('show-menu') && modelButton && modelButton !== exceptElement) {
+      this.closeVerbOrModelMenu(modelButton);
+    }
+    const verbDropdown = this.widget.querySelector('.verbs-container');
+    const verbButton = verbDropdown?.querySelector('.selected-verb');
+    if (verbDropdown && verbDropdown.classList.contains('show-menu') && verbButton && verbButton !== exceptElement) {
+      this.closeVerbOrModelMenu(verbButton);
+    }
+  }
+
+  updateAnalytics(verb) {
+    if (this.promptItems && this.promptItems.length > 0) {
+      this.promptItems.forEach((item) => {
+        const ariaLabel = item.getAttribute('aria-label') || '';
+        item.setAttribute('daa-ll', `${ariaLabel.slice(0, 20)}--${verb}--Prompt suggestion`);
+      });
+    }
+    if (this.genBtn) {
+      this.genBtn.setAttribute('daa-ll', `Generate--${verb}`);
+    }
+  }
+
+  clearSelectedModelState() {
+    this.selectedModelId = '';
+    this.selectedModelVersion = '';
+    this.selectedModelModule = '';
+    this.selectedModelText = '';
+  }
+
+  handleVerbLinkClick(link, verbList, selectedElement, menuIcon, inputPlaceHolder, modelList) {
+    return (e) => {
+      e.preventDefault();
+      e.stopPropagation();
+      const verbLinkTexts = [];
+      verbList.querySelectorAll('.verb-link').forEach((listLink) => {
+        listLink.parentElement.classList.remove('selected');
+        listLink.setAttribute('aria-selected', 'false');
+        const text = listLink.textContent.trim();
+        if (text) verbLinkTexts.push(text);
+      });
+      verbLinkTexts.sort((a, b) => b.length - a.length);
+      this.closeVerbOrModelMenu(selectedElement);
+      link.parentElement.classList.add('selected');
+      link.setAttribute('aria-selected', 'true');
+      if (modelList) {
+        this.selectedModelId = link.getAttribute('data-model-id');
+        this.selectedModelVersion = link.getAttribute('data-model-version');
+        this.selectedModelModule = link.getAttribute('data-model-module');
+        this.selectedModelText = link.textContent.trim();
+        const copiedNodes = link.cloneNode(true).childNodes;
+        copiedNodes[0].remove();
+        selectedElement.replaceChildren(...copiedNodes, menuIcon);
+        selectedElement.dataset.selectedModelId = this.selectedModelId;
+        selectedElement.dataset.selectedModelVersion = this.selectedModelVersion;
+      } else {
+        this.selectedVerbType = link.getAttribute('data-verb-type');
+        this.selectedVerbText = link.textContent.trim();
+        selectedElement.replaceChildren(this.selectedVerbText, menuIcon);
+        selectedElement.dataset.selectedVerb = this.selectedVerbType;
+      }
+      selectedElement.focus();
+      this.ensureSoundModuleLoaded();
+      const verbsWithoutPromptSuggestions = this.workflowCfg.targetCfg?.verbsWithoutPromptSuggestions ?? [];
+      if (!verbsWithoutPromptSuggestions.includes(this.selectedVerbType)) this.updateDropdownForVerb(this.selectedVerbType);
+      else this.widgetWrap.dispatchEvent(new CustomEvent('firefly-reinit-action-listeners'));
+      if (link.getAttribute('data-model-module') !== this.selectedVerbType) {
+        const oldModelContainer = this.widget.querySelector('.models-container');
+        const modelDropdown = this.modelDropdown();
+        if (oldModelContainer) {
+          if (modelDropdown.length > 1) {
+            const newModelContainer = createTag('div', { class: 'models-container', 'aria-label': 'Model options' });
+            newModelContainer.append(...modelDropdown);
+            oldModelContainer.replaceWith(newModelContainer);
+          } else {
+            oldModelContainer.remove();
+            this.clearSelectedModelState();
+          }
+        } else if (modelDropdown.length > 1) {
+          const actionContainer = this.widget.querySelector('.action-container');
+          if (actionContainer) {
+            const newModelContainer = createTag('div', { class: 'models-container', 'aria-label': 'Prompt options' });
+            newModelContainer.append(...modelDropdown);
+            actionContainer.append(newModelContainer);
+          }
+        } else this.clearSelectedModelState();
+      }
+      this.widgetWrap.setAttribute('data-selected-verb', this.selectedVerbType);
+      if (this.selectedModelId) this.widgetWrap.setAttribute('data-selected-model-id', this.selectedModelId);
+      else this.widgetWrap.removeAttribute('data-selected-model-id');
+      if (this.selectedModelVersion) this.widgetWrap.setAttribute('data-selected-model-version', this.selectedModelVersion);
+      else this.widgetWrap.removeAttribute('data-selected-model-version');
+      this.updateAnalytics(this.selectedVerbType);
+      if (this.genBtn) {
+        const img = this.genBtn.querySelector('img[src*=".svg"]');
+        this.genBtn.setAttribute(
+          'aria-label',
+          (this.genBtn.getAttribute('aria-label') || '').replace(
+            new RegExp(`\\b(${verbLinkTexts.join('|')})\\b`),
+            this.selectedVerbText,
+          ),
+        );
+        if (img) img.setAttribute('alt', `${this.genBtn.getAttribute('aria-label') || ''}`);
+      }
+    };
+  }
+
+  createDropdownItems(items, listContainer, selectedElement, menuIcon, inputPlaceHolder, isModelList) {
+    const fragment = document.createDocumentFragment();
+    items.forEach((item, idx) => {
+      const {
+        name,
+        type,
+        icon,
+        module,
+        id,
+        version,
+      } = item;
+      const listItem = createTag('li', {
+        class: 'verb-item',
+        role: 'presentation',
+      });
+      const selectedIcon = createTag('span', { class: 'selected-icon' }, '<svg><use xlink:href="#unity-checkmark-icon"></use></svg>');
+      const nameContainer = isModelList && createTag('span', { class: 'model-name' }, name.trim());
+      const link = createTag('a', {
+        href: '#',
+        class: isModelList ? 'verb-link model-link' : 'verb-link',
+        ...(!isModelList && { 'data-verb-type': type }),
+        ...(isModelList && { 'data-model-module': module }),
+        ...(isModelList && { 'data-model-id': id }),
+        ...(isModelList && { 'data-model-version': version }),
+        'aria-selected': 'false',
+        role: 'option',
+      }, `<img loading="lazy" src="${icon}" alt="" />${nameContainer ? nameContainer.outerHTML : name}`);
+      if (idx === 0) {
+        listItem.classList.add('selected');
+        link.setAttribute('aria-selected', 'true');
+      }
+      link.prepend(selectedIcon);
+      listItem.append(link);
+      fragment.append(listItem);
+    });
+    listContainer.append(fragment);
+    listContainer.addEventListener('click', (e) => {
+      const link = e.target.closest('.verb-link');
+      if (!link) return;
+      this.handleVerbLinkClick(link, listContainer, selectedElement, menuIcon, inputPlaceHolder, isModelList)(e);
+    });
+  }
+
+  modelDropdown() {
+    if (!this.hasModelOptions) return [];
+    const models = Array.isArray(this.models)
+      ? this.models.filter((obj) => obj.module === this.selectedVerbType)
+      : [];
+    if (!Array.isArray(models) || models.length === 0) return [];
+    const inputPlaceHolder = this.el.querySelector('.icon-placeholder-input').parentElement.textContent;
+    const selectedModelType = models[0].id;
+    const selectedModelVersion = models[0].version;
+    const selectedModelModule = models[0].module;
+    const nameContainer = createTag('span', { class: 'model-name' }, models[0].name.trim());
+    const selectedElement = createTag('button', {
+      class: 'selected-model',
+      'aria-expanded': 'false',
+      'aria-controls': 'model-menu',
+      'aria-label': 'model type',
+      'aria-haspopup': 'listbox',
+      role: 'combobox',
+      'aria-labelledby': 'listbox-label',
+      'data-selected-model-id': selectedModelType,
+      'data-selected-model-version': selectedModelVersion,
+      'data-selected-model-module': selectedModelModule,
+    }, `<img src="${models[0].icon}" alt="" />${nameContainer.outerHTML}`);
+    this.selectedModelModule = selectedModelModule;
+    this.selectedModelId = selectedModelType;
+    this.selectedModelVersion = selectedModelVersion;
+    this.widgetWrap.setAttribute('data-selected-model-id', this.selectedModelId);
+    this.widgetWrap.setAttribute('data-selected-model-version', this.selectedModelVersion);
+    this.widgetWrap.setAttribute('data-selected-verb', this.selectedVerbType);
+    this.selectedModelText = models[0].name.trim();
+    const menuIcon = createTag('span', { class: 'menu-icon' }, '<svg><use xlink:href="#unity-chevron-icon"></use></svg>');
+    const listItems = createTag('ul', { class: 'verb-list', id: 'model-menu', role: 'listbox', 'aria-labelledby': 'listbox-label' });
+    listItems.setAttribute('style', 'display: none;');
+    selectedElement.append(menuIcon);
+    const handleDocumentClick = (e) => {
+      const menuContainer = selectedElement.parentElement;
+      if (!menuContainer.contains(e.target)) {
+        document.removeEventListener('click', handleDocumentClick);
+        this.closeVerbOrModelMenu(selectedElement);
+      }
+    };
+    selectedElement.addEventListener('click', (e) => {
+      e.stopPropagation();
+      this.hidePromptDropdown(selectedElement);
+      this.showVerbMenu(selectedElement);
+      document.addEventListener('click', handleDocumentClick);
+    }, true);
+    selectedElement.addEventListener('keydown', (e) => {
+      if (e.key === 'Enter' || e.key === ' ') {
+        this.hidePromptDropdown(selectedElement);
+        this.showVerbMenu(selectedElement);
+      }
+      if (e.key === 'Escape' || e.code === 27) {
+        this.closeVerbOrModelMenu(selectedElement);
+        selectedElement.focus();
+      }
+    });
+    this.createDropdownItems(models, listItems, selectedElement, menuIcon, inputPlaceHolder, true);
+    return [selectedElement, listItems];
+  }
+
+  createActBtn(cfg, cls) {
+    if (!cfg) return null;
+    const txt = cfg.innerText?.trim();
+    const img = cfg.querySelector('img[src*=".svg"]');
+    if (img) img.setAttribute('alt', `${txt?.split('\n')[0]} ${this.selectedVerbText}`);
+    const btn = createTag('a', { href: '#', class: `unity-act-btn ${cls}`, 'daa-ll': `Generate--${this.selectedVerbType}`, 'aria-label': `${txt?.split('\n')[0]} ${this.selectedVerbText}` });
+    if (img) btn.append(createTag('div', { class: 'btn-ico' }, img));
+    if (txt) btn.append(createTag('div', { class: 'btn-txt' }, txt.split('\n')[0]));
+    this.genBtn = btn;
+    return btn;
+  }
+
+  async loadModels() {
+    const { origin } = window.location;
+    const baseUrl = (origin.includes('.aem.') || origin.includes('.hlx.'))
+      ? `https://main--unity--adobecom.${origin.includes('.hlx.') ? 'hlx' : 'aem'}.live`
+      : origin;
+    const modelFile = `${baseUrl}/unity/configs/prompt/model-picker.json`;
+    const results = await fetch(modelFile);
+    if (!results.ok) {
+      throw new Error('Failed to fetch models.');
+    }
+    const modelJson = await results.json();
+    this.models = modelJson?.content?.data;
+  }
+
+  async getModel() {
+    if (!this.hasModelOptions) return [];
+    try {
+      if (!this.models || Object.keys(this.models).length === 0) await this.loadModels();
+      return this.models;
+    } catch (e) {
+      window.lana?.log(`Message: Error loading models, Error: ${e}`, this.lanaOptions);
+      return [];
+    }
+  }
+
+  /**
+   * Groups model-picker rows by module (used by tests; marquee uses flat `this.models` from JSON).
+   * @param {Array<object>} data
+   * @returns {Record<string, Array<object>>}
+   */
+  createModelMap(data) {
+    const modelMap = {};
+    if (Array.isArray(data)) {
+      data.forEach((item) => {
+        if (item.type) {
+          if (!modelMap[item.module]) modelMap[item.module] = [];
+          modelMap[item.module].push({ name: item.name, id: item.id, version: item.version, icon: item.icon });
+        }
+      });
+    }
+    return modelMap;
+  }
+
+  /**
+   * Marquee subclass refreshes prompt suggestions; base is a no-op (style launcher has no suggestions).
+   */
+  async updateDropdownForVerb() {
+    await Promise.resolve();
+  }
+}
 
 const VIEWPORT_COL = { MOBILE: 0, TABLET: 1, DESKTOP: 2 };
 
@@ -200,21 +640,15 @@ function applyAuthoringPreviewGridLoadingPriorities(previewRows, selectedStyleId
 }
 
 /**
- * @param {*} widgetInstance — UnityWidget instance (workflow-firefly/widget.js)
+ * @param {*} widgetInstance — {@link UnityWidget} from this module
  * @param {{ styles: Array<{ picture: HTMLPictureElement, label: string, styleDescription: string, prompt: string }>, previewRows: Array<Array<HTMLPictureElement | null>> }} parsed
  */
 export async function mountPromptWithStyleSelectUI(widgetInstance, parsed) {
   const { styles, previewRows } = parsed;
   if (!styles.length) return;
 
-  const unityLibs = getUnityLibs();
-  /* Prompt + style-select sheet (card + `.unity-prompt-with-style-select` overrides). Shared primitives come from workflow priorityLoad of `workflow-firefly/widget.css`. */
-  await new Promise((resolve) => {
-    loadStyle(`${unityLibs}/core/widgets/prompt-with-style-select/prompt-with-style-select.css`, resolve);
-  });
-
   const { el } = widgetInstance;
-  /* verb-options: widget.css only gives models-container / action-container flex layout with this class */
+  /* verb-options: merged variant CSS (incl. former widget.css) enables models-container / action-container flex layout */
   const widgetWrap = createTag('div', { class: 'ex-unity-wrap verb-options' });
   const [widget, unitySprite] = ['ex-unity-widget', 'unity-sprite-container']
     .map((c) => createTag('div', { class: c }));
